Revert "fix: Fixes bugs and adds JavaDoc to the AST RBE."
This reverts commit 34cd01fcbf8395432222cf45188f1e24aeca6a19.

+++ src/org/jitsi/impl/neomedia/rtp/remotebitrateestimator/AimdRateControl.java
@@ -356,14 +356,14 @@ class AimdRateControl
         inExperiment = false;
     }
 
-    public void setEstimate(long bitrateBps, long nowMs)
+    public void setEstimate(int bitrateBps, long nowMs)
     {
         updated = true;
         bitrateIsInitialized = true;
         currentBitrateBps = changeBitrate(bitrateBps, bitrateBps, nowMs);
     }
 
-    public void setMinBitrate(long minBitrateBps)
+    public void setMinBitrate(int minBitrateBps)
     {
         minConfiguredBitrateBps = minBitrateBps;
         currentBitrateBps = Math.max(minBitrateBps, currentBitrateBps);
